Ensure github action runner is current
* Refactor runner install step into its own script
* On docker build, install the latest github actions/runner
* On startup, check that the downloaded runner is current
  - if not current, download the latest runner and reinstall

Under k8s, when running an outdated version of the action runner,
the application would detect that it was out of date and attempt
to upgrade itself, which would result in a restart of the pod running
the container. This would restart the runner once again with the old
version from the image, causing a crash loop backoff.

Signed-off-by: Gabrielle Madden <gabriellemadden@users.noreply.github.com>